Move pathlib implementation out of `__init__.py`
Using the `asyncio` package as a model.
